address code review feedback
