+++ Makefile
@@ -1,0 +1,2 @@
+format:
+	@swift format -i -r Package.swift Examples Sources
